my_following_podcast_pagination

+++ lib/core/constants/base_podcast_entitie/base_podcast_audio_info.dart
@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 
 class PodcastAudioInoEntitie extends Equatable {
-  final int podcastDuration;
+  final double podcastDuration;
   final String podcastUrl;
 
   const PodcastAudioInoEntitie(this.podcastDuration, this.podcastUrl);
